requirements.txt
Flask
Gunicorn
python-dotenv
pymysql
Flask-SQLAlchemy
